TT: Reduce limits issues w/ 12 bit immediates
This is a partial implementation...
Remove constraints on 12 bit immediates so overflowing the 12 bits
doesn't cause a compiler abort.  Instead, mask off the high bits.  And
with iadd, handle 16 bit values w/ a loadi.  Not sure if this is the
best implementation and doesn't (yet) handle unsigned, keep for now
revisit later.