TT: Multiple bug fixes for SSA, etc found w/ flag_checking=1
The smoke test started crashing when the wrapper was modified to
support non-imm op-codes.  The crash was in dead code: the live pass
had to change _lv to non_lv which no longer happens through the
wrapper.

To fix that, I enabled flag_checking=1 and founds/fixed lots of
issues.  However, the crash persists (likely related to setting
vuse).

Also, the liveness dbase was not updated in a stale code path, chasing
through assignments could bump into a statement w/o liveness set.