TT: Major renaming/cleanup
This came about because the namespace of the insns in the .md files
made it hard to be sure all rvtt insns were identified/matched to the
rvtt-insn list and tracked in the lookup map.  Fixing that namespace
led to the decision to put all of the relevant insns into the table
rather than matching multiple insns to a single entry.  It also begged
the question of the other functions/files being "riscv_sfpu" vs
"rvtt".  The new code is definitely different, I think in general
better, but I suspect the old names linger on in places, ugh.

The functional changes in this commit are:
 - the name matching for rtl insns
 - the register extraction/traversal code for rtl insns changed
   because the naming change picked up the compiler internal move as
   an sfpu insn and it follows a different pattern