TT: bug fixes for l1 reorder code (handle more rtl cases)
The compiler can optimize away an L1 load, store it to the stack, then
use the L1 intrinsic to load from the stack.  The chase back to find
the original load will find a, eg, mov insn.

Or, with test code of loading a pointer (not what it points to), the
compiler can chase back to the top of the fn and not find the load
since it was in a parameter