[pr13] Use a graph walk for SFPNOP insertion
We can't just look at the immediate successor blocks, but must walk
the graph until we find a reachable dependent insn. Obviously we need
to be careful of loops -- hence marking the visited bbs.  Note that
the first BB might itself be reachable, and we want to check it has no
use of the result. If we reach a non-dependent SPU insn, we do not
need a NOP.